#7 Introduce rate normalization support

+++ services/performance/src/main/java/org/jage/performance/node/category/PerformanceRate.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @AllArgsConstructor
 @ToString
 public class PerformanceRate {
 
     @Getter
-    private final int rate;
+    private final BigInteger rate;
 
 }
